+++ pybop/costs/base_cost.py
@@ -96,13 +96,9 @@ class BaseCost:
         """
         # Note, we use the transformation and parameter properties here to enable
         # differing attributes within the `LogPosterior` class
-
-        # Apply transformation if needed
         self.has_transform = self.transformation is not None and apply_transform
-        if self.has_transform:
-            model_inputs = self.transformation.to_model(inputs)
-        else:
-            model_inputs = inputs
+        model_inputs = self.parameters.verify(self._apply_transformations(inputs))
+        self.parameters.update(values=list(model_inputs.values()))
 
         # Check whether we are maximising or minimising via:
         # | `minimising` | `self.minimising` | `for_optimiser` |
@@ -112,10 +108,6 @@ class BaseCost:
         # | `False`      | `False`           | `True`          |
         # | `True`       | `False`           | `False`         |
         minimising = self.minimising or not for_optimiser
-
-        # Validate inputs, update parameters
-        model_inputs = self.parameters.verify(model_inputs)
-        self.parameters.update(values=list(model_inputs.values()))
 
         y, dy = None, None
         if self._has_separable_problem:
@@ -136,6 +128,10 @@ class BaseCost:
         return self.compute(y, dy=dy, calculate_grad=calculate_grad) * (
             1 if minimising else -1
         )
+
+    def _apply_transformations(self, inputs):
+        """Apply transformation if needed"""
+        return self.transformation.to_model(inputs) if self.has_transform else inputs
 
     def compute(self, y: dict, dy: ndarray, calculate_grad: bool = False):
         """