+++ pybop/optimisers/base_pints_optimiser.py
@@ -11,7 +11,14 @@ from pints import RectangularBoundaries as PintsRectangularBoundaries
 from pints import SequentialEvaluator as PintsSequentialEvaluator
 from pints import strfloat as PintsStrFloat
 
-from pybop import AdamWImpl, BaseOptimiser, GradientDescentImpl, OptimisationResult
+from pybop import (
+    AdamWImpl,
+    BaseJaxCost,
+    BaseOptimiser,
+    GradientDescentImpl,
+    OptimisationResult,
+    SequentialJaxEvaluator,
+)
 
 
 class BasePintsOptimiser(BaseOptimiser):
@@ -201,17 +208,20 @@ class BasePintsOptimiser(BaseOptimiser):
             return self.cost_call(x, calculate_grad=self._needs_sensitivities)
 
         # Create evaluator object
-        if self._parallel:
-            # Get number of workers
-            n_workers = self._n_workers
-
-            # For population based optimisers, don't use more workers than
-            # particles!
-            if isinstance(self.optimiser, PintsPopulationBasedOptimiser):
-                n_workers = min(n_workers, self.optimiser.population_size())
-            evaluator = PintsParallelEvaluator(fun, n_workers=n_workers)
+        if isinstance(self.cost, BaseJaxCost):
+            evaluator = SequentialJaxEvaluator(fun)
         else:
-            evaluator = PintsSequentialEvaluator(fun)
+            if self._parallel:
+                # Get number of workers
+                n_workers = self._n_workers
+
+                # For population based optimisers, don't use more workers than
+                # particles!
+                if isinstance(self.optimiser, PintsPopulationBasedOptimiser):
+                    n_workers = min(n_workers, self.optimiser.population_size())
+                evaluator = PintsParallelEvaluator(fun, n_workers=n_workers)
+            else:
+                evaluator = PintsSequentialEvaluator(fun)
 
         # Keep track of current best and best-guess scores.
         fb = fg = np.inf