+++ pybop/optimisers/scipy_optimisers.py
@@ -5,7 +5,7 @@ from typing import Union
 import numpy as np
 from scipy.optimize import Bounds, OptimizeResult, differential_evolution, minimize
 
-from pybop import BaseOptimiser, OptimisationResult
+from pybop import BaseOptimiser, OptimisationResult, SciPyEvaluator
 
 
 class BaseSciPyOptimiser(BaseOptimiser):
@@ -77,6 +77,14 @@ class BaseSciPyOptimiser(BaseOptimiser):
         result : pybop.Result
             The result of the optimisation including the optimised parameter values and cost.
         """
+
+        # Choose method to evaluate
+        def fun(x):
+            return self.cost_call(x, calculate_grad=self._needs_sensitivities)
+
+        # Create evaluator object
+        self.evaluator = SciPyEvaluator(fun)
+
         # Run with timing
         start_time = time()
         result = self._run_optimiser()
@@ -191,12 +199,15 @@ class SciPyMinimize(BaseSciPyOptimiser):
                 # Nest this option within an options dictionary for SciPy minimize
                 self._options["options"]["maxiter"] = self.unset_options.pop(key)
 
+        if self._options["jac"] is True:
+            self._needs_sensitivities = True
+
     def cost_wrapper(self, x):
         """
         Scale the cost function, preserving the sign convention, and eliminate nan values
         """
         if not self._options["jac"]:
-            cost = self.cost_call(x)
+            cost = self.evaluator.evaluate(x)
             self.log_update(x=[x], cost=cost)
             scaled_cost = cost / self._cost0
             if np.isinf(scaled_cost):
@@ -206,7 +217,7 @@ class SciPyMinimize(BaseSciPyOptimiser):
                 )  # for fake finite gradient
             return scaled_cost
 
-        L, dl = self.cost_call(x, calculate_grad=True)
+        L, dl = self.evaluator.evaluate(x)
         self.log_update(x=[x], cost=L)
         return (L / self._cost0, dl / self._cost0)
 
@@ -234,7 +245,8 @@ class SciPyMinimize(BaseSciPyOptimiser):
                 cost_best = intermediate_result.fun * self._cost0
             else:
                 x_best = intermediate_result
-                cost_best = self.cost_call(x_best)
+                result = self.evaluator.evaluate(x_best)
+                cost_best = result[0] if self._needs_sensitivities else result
 
             self.log_update(x_best=x_best, cost_best=cost_best)
 
@@ -245,7 +257,8 @@ class SciPyMinimize(BaseSciPyOptimiser):
         )
 
         # Compute the absolute initial cost and resample if required
-        self._cost0 = np.abs(self.cost_call(self.x0))
+        result = self.evaluator.evaluate(self.x0)
+        self._cost0 = np.abs(result[0] if self._needs_sensitivities else result)
         if np.isinf(self._cost0):
             for _i in range(1, self.num_resamples):
                 try:
@@ -257,7 +270,8 @@ class SciPyMinimize(BaseSciPyOptimiser):
                         stacklevel=2,
                     )
                     break
-                self._cost0 = np.abs(self.cost_call(self.x0))
+                result = self.evaluator.evaluate(self.x0)
+                self._cost0 = np.abs(result[0] if self._needs_sensitivities else result)
                 if not np.isinf(self._cost0):
                     break
             if np.isinf(self._cost0):
@@ -427,7 +441,7 @@ class SciPyDifferentialEvolution(BaseSciPyOptimiser):
             )
 
         def cost_wrapper(x):
-            cost = self.cost_call(x)
+            cost = self.evaluator.evaluate(x)
             self.log_update(x=[x], cost=cost)
             return cost