+++ pybop/problems/base_problem.py
@@ -62,17 +62,19 @@ class BaseProblem:
         self.parameters = parameters
         self.parameters.reset_initial_value()
 
-        self._model = model
+        self._model = model.copy() if model is not None else None
         self.eis = False
         self.domain = "Time [s]"
         self.check_model = check_model
         self.signal = signal or ["Voltage [V]"]
         self.set_initial_state(initial_state)
         self._dataset = None
-        self._domain_data = None
         self._target = None
         self.verbose = False
         self.failure_output = np.asarray([np.inf])
+        self.exception = [
+            "These parameter values are infeasible."
+        ]  # TODO: Update to a utility function and add to it on exception creation
         if isinstance(self._model, BaseModel):
             self.eis = self.model.eis
             self.domain = "Frequency [Hz]" if self.eis else "Time [s]"